get 11ty beforeBuild working
