+++ cli/comicgeeks/esm/index.mjs
@@ -1,0 +1,16 @@
+import ComicGeeks from '../dist/index.js';
+
+export default ComicGeeks;
+
+export const {
+    fetchCollection,
+    fetchWishList,
+    fetchReleases,
+    fetchPulls,
+    fetchSearchResults,
+    fetchUser,
+    FilterTypes,
+    SortTypes,
+    CollectionTypes,
+    Publishers
+} = ComicGeeks;
