+++ cli/comicgeeks/typings/index.d.ts
@@ -1,0 +1,3 @@
+export * from './lib/constants';
+export * from './lib/interfaces';
+export * from './lib/methods';
