--- comicgeeks/esm/index.mjs
@@ -1,16 +1,0 @@
-import ComicGeeks from '../dist/index.js';
-
-export default ComicGeeks;
-
-export const {
-    fetchCollection,
-    fetchWishList,
-    fetchReleases,
-    fetchPulls,
-    fetchSearchResults,
-    fetchUser,
-    FilterTypes,
-    SortTypes,
-    CollectionTypes,
-    Publishers
-} = ComicGeeks;
