--- comicgeeks/typings/index.d.ts
@@ -1,3 +1,0 @@
-export * from './lib/constants';
-export * from './lib/interfaces';
-export * from './lib/methods';
