--- comicgeeks/typings/lib/sorts.d.ts
@@ -1,4 +1,0 @@
-import type { Comic } from './interfaces';
-declare type Sorter = 'Alpha' | 'Community' | 'PotW' | 'Price';
-declare const _default: Record<Sorter, (a: Comic, b: Comic) => number>;
-export default _default;
